Вынес работу с comments в модуль render-comments
