+++ js/show-full-size-mode.js
@@ -1,3 +1,5 @@
+import { renderComments } from './render-comments.js';
+
 const MAXIMUM_COUNT_COMMENTS_SHOWN = 5;
 
 const bigPicture = document.querySelector('.big-picture');
@@ -6,42 +8,11 @@ const fullSizeModeCloseElement = bigPicture.querySelector(
 );
 const bigPictureImage = bigPicture.querySelector('.big-picture__img img');
 const likesCount = bigPicture.querySelector('.likes-count');
-const socialCommentShownCount = bigPicture.querySelector(
-  '.social__comment-shown-count'
-);
-const socialCommentTotalCount = bigPicture.querySelector(
-  '.social__comment-total-count'
-);
 const socialCaption = bigPicture.querySelector('.social__caption');
-const commentsList = bigPicture.querySelector('.social__comments');
 const commentsLoader = bigPicture.querySelector('.comments-loader');
-const commentTemplate = document
-  .querySelector('#comment')
-  .content.querySelector('.social__comment');
 
 let onCommentsLoaderShowMore = null;
 
-const getCommentElement = ({ avatar, name, message }) => {
-  const commentElement = commentTemplate.cloneNode(true);
-  const avatarElement = commentElement.querySelector('.social__picture');
-
-  avatarElement.src = avatar;
-  avatarElement.alt = name;
-  commentElement.querySelector('.social__text').textContent = message;
-  return commentElement;
-};
-// TODO: вынести в отдельный модуль
-const renderComments = (count, arrayComments) => {
-  const comments = arrayComments.slice(0, count);
-  const isLengthsMatch = comments.length === arrayComments.length;
-
-  commentsList.textContent = '';
-  commentsList.append(...comments.map(getCommentElement));
-
-  socialCommentShownCount.textContent = comments.length;
-  socialCommentTotalCount.textContent = arrayComments.length;
-  commentsLoader.classList[isLengthsMatch ? 'add' : 'remove']('hidden');
-};
 const handlerShowMore = (shownCount, comments) => () => {
   shownCount += MAXIMUM_COUNT_COMMENTS_SHOWN;
   renderComments(shownCount, comments);
